Outsource SeederService a Create SeederController
