+++ backend/SmartEdu/Services/AccountService/AccountService.cs
@@ -582,855 +582,6 @@ namespace SmartEdu.Services.AccountService
 
         }
 
-        public async Task<ServerResponse<object>> SeedingUsers()
-        {
-            var serverResponse = new ServerResponse<object>();
-
-            var roles = new List<string> { "User" };
-
-            var registerUserDTOs =new List<RegisterUserDTO>
-            {
-                new RegisterUserDTO
-                {
-                    FullName = "Nguyen Thi Giang",
-                    UserName = "giang",
-                    Email = "giang@c3chuvanan.edu.vn",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 3 
-                },
-                new RegisterUserDTO
-                {
-                    FullName = "Pham Thi Nguyet Anh",
-                    UserName = "anhpham",
-                    Email = "anh.pham@c3chuvanan.edu.vn",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 3
-                },
-                new RegisterUserDTO
-                {
-                    FullName = "Nguyen Dam Thuy Duong",
-                    UserName = "duong",
-                    Email = "duong@c3chuvanan.edu.vn",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 3
-                },
-                new RegisterUserDTO
-                {
-                    FullName = "Nguyen Thi Huong",
-                    UserName = "huong",
-                    Email = "nguyenhuongg1104@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = new List<string> { "Admin" },
-                    Type = 0
-                },
-                new RegisterUserDTO
-                {
-                    FullName = "Trinh Van Phuc",
-                    UserName = "phuc",
-                    Email = "phuctv1112004@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = new List<string> { "Admin" },
-                    Type = 0
-                },
-                new RegisterUserDTO // 1
-                {
-                    FullName = "Leu Ngoc An",
-                    UserName = "an0149107377",
-                    Email = "an0149107377@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 1
-                },
-                new RegisterUserDTO // 2
-                {
-                    FullName = "Nguyen Ngoc Tram Anh",
-                    UserName = "anh0116701758",
-                    Email = "anh0116701758@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 1
-                },
-                new RegisterUserDTO // 3
-                {
-                    FullName = "Pham Viet Anh",
-                    UserName = "anh0133530105",
-                    Email = "anh0133530105@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 1
-                },
-                new RegisterUserDTO // 4
-                {
-                    FullName = "Bui Thi Quynh Anh",
-                    UserName = "anh3616460199",
-                    Email = "anh3616460199@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 1
-                },
-                new RegisterUserDTO // 5
-                {
-                    FullName = "Vu Duc Anh",
-                    UserName = "anh1734705112",
-                    Email = "anh1734705112@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 1
-                },
-                new RegisterUserDTO // 6
-                {
-                    FullName = "Nguyen Phung Linh Chi",
-                    UserName = "chi0149107460",
-                    Email = "chi0149107460@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 1
-                },
-                new RegisterUserDTO // 7
-                {
-                    FullName = "Duong My Dung",
-                    UserName = "dung0116989583",
-                    Email = "dung0116989583@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 1
-                },
-                new RegisterUserDTO // 8
-                {
-                    FullName = "Nguyen Manh Duy",
-                    UserName = "duy0150499677",
-                    Email = "duy0150499677@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 1
-                },
-                new RegisterUserDTO // 9
-                {
-                    FullName = "Pham Phuong Duy",
-                    UserName = "duy0116479556",
-                    Email = "duy0116479556@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 1
-                },
-                new RegisterUserDTO // 10
-                {
-                    FullName = "Nguyen Thuy Duong",
-                    UserName = "duong0133058338",
-                    Email = "duong0133058338@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 1
-                },
-                new RegisterUserDTO // 11
-                {
-                    FullName = "Luu Minh Hang",
-                    UserName = "hang0116971907",
-                    Email = "hang0116971907@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 1
-                },
-                new RegisterUserDTO // 12
-                {
-                    FullName = "Nguyen Huu Minh Hoang",
-                    UserName = "hoang0116952467",
-                    Email = "hoang0116952467@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 1
-                },
-                new RegisterUserDTO // 13
-                {
-                    FullName = "Nguyen Huy Hoang",
-                    UserName = "hoang0149107469",
-                    Email = "hoang0149107469@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 1
-                },
-                new RegisterUserDTO // 14
-                {
-                    FullName = "Nguyen Duc Huy",
-                    UserName = "huy0116575041",
-                    Email = "huy0116575041@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 1
-                },
-                new RegisterUserDTO // 15
-                {
-                    FullName = "Vu Duc Huy",
-                    UserName = "huy0116701214",
-                    Email = "huy0116701214@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 1
-                },
-                new RegisterUserDTO // 16
-                {
-                    FullName = "Nguyen Trung Kien",
-                    UserName = "kien3316697828",
-                    Email = "kien3316697828@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 1
-                },
-                new RegisterUserDTO // 17
-                {
-                    FullName = "Le Duy Khiem",
-                    UserName = "khiem0137566271",
-                    Email = "khiem0137566271@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 1
-                },
-                new RegisterUserDTO // 18
-                {
-                    FullName = "Nguyen Minh Khue",
-                    UserName = "khue0150707773",
-                    Email = "khue0150707773@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 1
-                },
-                new RegisterUserDTO //19
-                {
-                    FullName = "Bui Hai Lam",
-                    UserName = "lam0134001874",
-                    Email = "lam0134001874@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 1
-                },
-                new RegisterUserDTO //20
-                {
-                    FullName = "Nguyen Ha Gia Linh",
-                    UserName = "linh0116973081",
-                    Email = "linh0116973081@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 1
-                },
-                new RegisterUserDTO //21
-                {
-                    FullName = "Nguyen Phuc Loc",
-                    UserName = "loc0116574901",
-                    Email = "loc0116574901@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 1
-                },
-                new RegisterUserDTO //22
-                {
-                    FullName = "Hoang Huong Ly",
-                    UserName = "ly0150498659",
-                    Email = "ly0150498659@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 1
-                },
-                new RegisterUserDTO //23
-                {
-                    FullName = "Trinh Xuan Minh",
-                    UserName = "minh0155210714",
-                    Email = "minh0155210714@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 1
-                },
-                new RegisterUserDTO //24
-                {
-                    FullName = "Nguyen Vu Tien Nam",
-                    UserName = "nam0116695176",
-                    Email = "nam0116695176@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 1
-                },
-                new RegisterUserDTO //25
-                {
-                    FullName = "Pham Thi Hang Nga",
-                    UserName = "nga0134705129",
-                    Email = "nga0134705129@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 1
-                },
-                new RegisterUserDTO //26
-                {
-                    FullName = "Hoang Kim Ngan",
-                    UserName = "ngan0133530178",
-                    Email = "ngan0133530178@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 1
-                },
-                new RegisterUserDTO //27
-                {
-                    FullName = "Hoang Yen Nhi",
-                    UserName = "nhi0134705133",
-                    Email = "nhi0134705133@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 1
-                },
-                new RegisterUserDTO //28
-                {
-                    FullName = "Mai Ngo Thien Phu",
-                    UserName = "phu0134705135",
-                    Email = "phu0134705135@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 1
-                },
-                new RegisterUserDTO //29
-                {
-                    FullName = "Le Tien Tam",
-                    UserName = "tam0100712147",
-                    Email = "tam0100712147@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 1
-                },
-                new RegisterUserDTO //30
-                {
-                    FullName = "Tran Xuan Toan",
-                    UserName = "toan0133487571",
-                    Email = "toan0133487571@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 1
-                },new RegisterUserDTO //31
-                {
-                    FullName = "Le Minh Tuan",
-                    UserName = "tuan0133530254",
-                    Email = "tuan0133530254@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 1
-                },
-                new RegisterUserDTO //32
-                {
-                    FullName = "Duong Tat Thanh",
-                    UserName = "thanh0134100841",
-                    Email = "thanh0134100841@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 1
-                },
-                new RegisterUserDTO //33
-                {
-                    FullName = "Le Cao Thang",
-                    UserName = "thang0155203520",
-                    Email = "thang0155203520@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 1
-                },
-                new RegisterUserDTO //34
-                {
-                    FullName = "Dang Phuong Thuy",
-                    UserName = "thuy0144525596",
-                    Email = "thuy0144525596@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 1
-                },
-                new RegisterUserDTO //35
-                {
-                    FullName = "Nguyen Thu Trang",
-                    UserName = "trang0134100853",
-                    Email = "trang0134100853@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 1
-                },
-                new RegisterUserDTO //36
-                {
-                    FullName = "Nguyen Thuy Trang",
-                    UserName = "trang0134100856",
-                    Email = "trang0134100856@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 1
-                },
-
-                // Parents
-                new RegisterUserDTO // 1
-                {
-                    FullName = "Leu Minh Duc",
-                    UserName = "duc0149107377",
-                    Email = "duc0149107377@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 2
-                },
-                new RegisterUserDTO // 2
-                {
-                    FullName = "Nguyen Ngoc Bao",
-                    UserName = "bao0116701758",
-                    Email = "bao0116701758@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 2
-                },
-                new RegisterUserDTO // 3
-                {
-                    FullName = "Pham Viet Hung",
-                    UserName = "hung0133530105",
-                    Email = "hung0133530105@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 2
-                },
-                new RegisterUserDTO // 4
-                {
-                    FullName = "Bui Thi Quynh Mai",
-                    UserName = "mai3616460199",
-                    Email = "mai3616460199@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 2
-                },
-                new RegisterUserDTO // 5
-                {
-                    FullName = "Vu Tuan Minh ",
-                    UserName = "minh1734705112",
-                    Email = "minh1734705112@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 2
-                },
-                new RegisterUserDTO // 6
-                {
-                    FullName = "Nguyen The Toan",
-                    UserName = "toan0149107460",
-                    Email = "toan0149107460@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 2
-                },
-                new RegisterUserDTO // 7
-                {
-                    FullName = "Duong My Hanh",
-                    UserName = "hanh0116989583",
-                    Email = "hanh0116989583@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 2
-                },
-                new RegisterUserDTO // 8
-                {
-                    FullName = "Nguyen Manh Dung",
-                    UserName = "dung0150499677",
-                    Email = "dung0150499677@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 2
-                },
-                new RegisterUserDTO // 9
-                {
-                    FullName = "Pham Phuong Thuy",
-                    UserName = "thuy0116479556",
-                    Email = "thuy0116479556@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 2
-                },
-                new RegisterUserDTO // 10
-                {
-                    FullName = "Nguyen Thuy Dung",
-                    UserName = "dung0133058338",
-                    Email = "dung0133058338@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 2
-                },
-                new RegisterUserDTO // 11
-                {
-                    FullName = "Luu Thi Hong",
-                    UserName = "hong0116971907",
-                    Email = "hong0116971907@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 2
-                },
-                new RegisterUserDTO // 12
-                {
-                    FullName = "Nguyen Huu Minh",
-                    UserName = "minh0116952467",
-                    Email = "minh0116952467@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 2
-                },
-                new RegisterUserDTO // 13
-                {
-                    FullName = "Nguyen Huy Long",
-                    UserName = "long0149107469",
-                    Email = "long0149107469@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 2
-                },
-                new RegisterUserDTO // 14
-                {
-                    FullName = "Nguyen Duc Toan",
-                    UserName = "toan0116575041",
-                    Email = "toan0116575041@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 2
-                },
-                new RegisterUserDTO // 15
-                {
-                    FullName = "Vu Duc Dam",
-                    UserName = "dam0116701214",
-                    Email = "dam0116701214@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 2
-                },
-                new RegisterUserDTO // 16
-                {
-                    FullName = "Nguyen Trung Thanh",
-                    UserName = "thanh3316697828",
-                    Email = "thanh3316697828@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 2
-                },
-                new RegisterUserDTO // 17
-                {
-                    FullName = "Le Duy Khanh",
-                    UserName = "khanh0137566271",
-                    Email = "khanh0137566271@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 2
-                },
-                new RegisterUserDTO // 18
-                {
-                    FullName = "Nguyen Minh Ngoc",
-                    UserName = "ngoc0150707773",
-                    Email = "ngoc0150707773@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 2
-                },
-                new RegisterUserDTO //19
-                {
-                    FullName = "Bui Hai Luong",
-                    UserName = "luong0134001874",
-                    Email = "luong0134001874@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 2
-                },
-                new RegisterUserDTO //20
-                {
-                    FullName = "Nguyen Gia Bao",
-                    UserName = "bao0116973081",
-                    Email = "bao0116973081@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 2
-                },
-                new RegisterUserDTO //21
-                {
-                    FullName = "Nguyen Phuc Duc",
-                    UserName = "duc0116574901",
-                    Email = "duc0116574901@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 2
-                },
-                new RegisterUserDTO //22
-                {
-                    FullName = "Hoang Huong Nhung",
-                    UserName = "nhung0150498659",
-                    Email = "nhung0150498659@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 2
-                },
-                new RegisterUserDTO //23
-                {
-                    FullName = "Trinh Xuan Thanh",
-                    UserName = "thanh0155210714",
-                    Email = "thanh0155210714@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 2
-                },
-                new RegisterUserDTO //24
-                {
-                    FullName = "Nguyen Vu Tien",
-                    UserName = "tien0116695176",
-                    Email = "tien0116695176@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 2
-                },
-                new RegisterUserDTO //25
-                {
-                    FullName = "Pham Thi Hang",
-                    UserName = "hang0134705129",
-                    Email = "hang0134705129@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 2
-                },
-                new RegisterUserDTO //26
-                {
-                    FullName = "Hoang Kim Trang",
-                    UserName = "trang0133530178",
-                    Email = "trang0133530178@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 2
-                },
-                new RegisterUserDTO //27
-                {
-                    FullName = "Hoang Thi Men",
-                    UserName = "men0134705133",
-                    Email = "men0134705133@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 2
-                },
-                new RegisterUserDTO //28
-                {
-                    FullName = "Mai Thien Hoang",
-                    UserName = "hoang0134705135",
-                    Email = "hoang0134705135@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 2
-                },
-                new RegisterUserDTO //29
-                {
-                    FullName = "Le Tien Tuan",
-                    UserName = "tuan0100712147",
-                    Email = "tuan0100712147@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 2
-                },
-                new RegisterUserDTO //30
-                {
-                    FullName = "Tran Xuan Linh",
-                    UserName = "linh0133487571",
-                    Email = "linh0133487571@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 2
-                },new RegisterUserDTO //31
-                {
-                    FullName = "Le Minh Quan",
-                    UserName = "quan0133530254",
-                    Email = "quan0133530254@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 2
-                },
-                new RegisterUserDTO //32
-                {
-                    FullName = "Duong Tat Phuc",
-                    UserName = "phuc0134100841",
-                    Email = "phuc0134100841@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 2
-                },
-                new RegisterUserDTO //33
-                {
-                    FullName = "Le Cao Duc",
-                    UserName = "duc0155203520",
-                    Email = "duc0155203520@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 2
-                },
-                new RegisterUserDTO //34
-                {
-                    FullName = "Dang Phuong Linh",
-                    UserName = "linh0144525596",
-                    Email = "linh0144525596@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 2
-                },
-                new RegisterUserDTO //35
-                {
-                    FullName = "Nguyen Le Thu",
-                    UserName = "thu0134100853",
-                    Email = "thu0134100853@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 2
-                },
-                new RegisterUserDTO //36
-                {
-                    FullName = "Nguyen  Thi Thuy",
-                    UserName = "thuy0134100856",
-                    Email = "thuy0134100856@gmail.com",
-                    Password = "Sm@rtEdu1",
-                    Roles = roles,
-                    Type = 2
-                },
-            };
-
-            foreach (var registerUserDTO in registerUserDTOs)
-            {
-                var user = _mapper.Map<User>(registerUserDTO);
-                // tao tai khoan cho mooi giao vien va them vao trong DB
-
-                await _userManager.CreateAsync(user,
-                    registerUserDTO.Password);
-                await _userManager.AddToRolesAsync(user,
-                    roles);
-            }
-
-            serverResponse.Message = "Seeding users successfully";
-            return serverResponse;
-        }
-
-        public async Task<ServerResponse<object>> SeedingTeachers()
-        {
-            var serverResponse = new ServerResponse<object>();
-
-            var users = await _userManager.Users.ToListAsync();
-
-            Teacher teacher = null;
-
-            //for (var i = 0; i < users.Count; i++)
-            //{
-            //    var user = users.ElementAt(i);
-            //    if (user.Type == 3)
-            //    {
-            //        teacher = new Teacher
-            //        {
-            //            Id = 1,
-            //            UserId = user.Id,
-            //            MainClassId = i + 1,
-            //            SubjectId = 1
-            //        };
-            //    }
-            //}
-
-            var mainClassId = 1;
-            var id = 1;
-            var teachers = new List<Teacher>();
-            foreach (var user in users)
-            {
-                if (user.Type == 3)
-                {
-                    var t = new Teacher
-                    {
-                        UserId = user.Id,
-                        MainClassId = mainClassId,
-                        SubjectId = 1
-                    };
-                    teachers.Add(t);
-                    id++;
-                    mainClassId++;
-                }
-                
-            }
-            await _unitOfWork.TeacherRepository.AddRange(teachers);
-            
-            await _unitOfWork.Save();
-
-            serverResponse.Message = "Seeding Teachers successfully";
-
-            return serverResponse;
-        }
-
-        public async Task<ServerResponse<object>> SeedingStudents()
-        {
-            var serverResponse = new ServerResponse<object>();
-
-            var users = await _userManager.Users.ToListAsync();
-            var students = new List<Student>();
-            var id = 1;
-            var count = 0;
-            var mainClassId = 1;
-
-            foreach (var user in users)
-            {
-                if (user.Type == 1)
-                {
-
-                    if (count == 36)
-                    {
-                        mainClassId++;
-                        count = 0;
-                    }
-                    var student = new Student
-                    {
-                        UserId = user.Id,
-                        ParentId = id,
-                        MainClassId = mainClassId,
-                    };
-                    students.Add(student);
-                    count++;
-                    id++;
-                }
-            }
-
-            await _unitOfWork.StudentRepository.AddRange(students);
-            await _unitOfWork.Save();
-
-            serverResponse.Message = "Add students successfully";
-            return serverResponse;
-        }
-
-        public async Task<ServerResponse<object>> SeedingParents()
-        {
-            var serverResponse = new ServerResponse<object>();
-
-            var users = await _userManager.Users.ToListAsync();
-
-            var parents = new List<Parent>();
-
-            var id = 1;
-            
-            foreach(var user in users)
-            {
-                if (user.Type == 2)
-                {
-                    var parent = new Parent
-                    {
-                        UserId = user.Id,
-                    };
-                    parents.Add(parent);
-                    id++;
-                }
-                
-            }
-            
-            await _unitOfWork.ParentRepository.AddRange(parents);
-
-            await _unitOfWork.Save();
-            serverResponse.Message = "Seeding parents successfully.";
-            return serverResponse;
-        }
-
-        public async Task<ServerResponse<object>> SeedingData()
-        {
-            await SeedingUsers();
-            await SeedingTeachers();
-            await SeedingParents();
-            await SeedingStudents();
-
-            var serverResponse = new ServerResponse<object>();
-            serverResponse.Message = "Seeding data successfully";
-            return serverResponse;
-        }
+        
     }
 }
