+++ backend/SmartEdu/Services/AccountService/IAccountService.cs
@@ -29,10 +29,6 @@ namespace SmartEdu.Services.AccountService
         Task<ServerResponse<AuthResponseUserDTO>> FacebookLogin(FacebookAuthUserDTO facebookAuthUserDTO);
         Task<ServerResponse<GetUserDTO>> UpdateProfileImage(MultipleFilesModel model, ModelStateDictionary modelState);
 
-        Task<ServerResponse<object>> SeedingUsers();
-        Task<ServerResponse<object>> SeedingTeachers();
-        Task<ServerResponse<object>> SeedingStudents();
-        Task<ServerResponse<object>> SeedingParents();
-        Task<ServerResponse<object>> SeedingData();
+        
     }
 }
